Admin Logs, Image Handler Fix, Multi Language Fix.

+++ app/Helpers/ImageHelper.php
@@ -1,12 +1,15 @@
 <?php
 namespace App\Helpers;
 
-use DataTables;
 use Intervention\Image\Facades\Image;
 
 class ImageHelper
 {
     public static function uploadImage($file, $fileName = null) {
+        if(empty($file)){
+            return null;
+        }
+
         $image = $file;
         if(empty($fileName)){
             $image_name = time() . '.' . $file->getClientOriginalExtension();
@@ -22,12 +25,21 @@ class ImageHelper
     }
 
     public static function removeImage($path){
-        if(file_exists(public_path($path))){
+        if(file_exists(public_path($path)) && !empty($path)){
             unlink(public_path($path));
         }
     }
 
-    public static function updateImage($file, $path, $fileName = null){
+    public static function updateImage($file, $path, $is_remove = "false", $fileName = null){
+        if($is_remove == "true"){
+            self::removeImage($path);
+            return null;
+        }
+
+        if(empty($file)){
+            return $path;
+        }
+
         self::removeImage($path);
         return self::uploadImage($file, $fileName);
     }
